+++ src/test/java/com/ruby/mall/controller/CsrfTest.java
@@ -2,6 +2,8 @@ package com.ruby.mall.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ruby.mall.common.CheckResult;
+import com.ruby.mall.dto.BuyItem;
+import com.ruby.mall.dto.CreateOrderRequest;
 import com.ruby.mall.dto.UserRegisterRequest;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,6 +13,9 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.RequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.httpBasic;
 
@@ -32,6 +37,27 @@ public class CsrfTest {
         RequestBuilder requestBuilder = MockMvcRequestBuilders
                 .post("/admin/register")
                 .with(httpBasic("admin@gmail.com", "admin"))
+                .contentType(MediaType.APPLICATION_JSON)
+                .content(json);
+
+        new CheckResult(mockMvc).check(requestBuilder, 403);
+    }
+
+    @Test
+    public void createOrder_noCsrf() throws Exception {
+        CreateOrderRequest createOrderRequest = new CreateOrderRequest();
+        List<BuyItem> buyItemList = new ArrayList<>();
+
+        BuyItem buyItem1 = new BuyItem();
+        buyItem1.setProductId(100);
+        buyItem1.setQuantity(1);
+        buyItemList.add(buyItem1);
+
+        createOrderRequest.setOrderList(buyItemList);
+        String json = objectMapper.writeValueAsString(createOrderRequest);
+        RequestBuilder requestBuilder = MockMvcRequestBuilders
+                .post("/users/{userId}/orders", 1)
+                .with(httpBasic("user2@gmail.com", "user2"))
                 .contentType(MediaType.APPLICATION_JSON)
                 .content(json);
