+++ src/test/java/com/ruby/mall/controller/OrderControllerTest.java
@@ -144,27 +144,6 @@ public class OrderControllerTest {
         new CheckResult(mockMvc).check(requestBuilder, statusCode.getResponseCode(), statusCode.getResponseBody());
     }
 
-    @Test
-    public void createOrder_noCsrf() throws Exception {
-        CreateOrderRequest createOrderRequest = new CreateOrderRequest();
-        List<BuyItem> buyItemList = new ArrayList<>();
-
-        BuyItem buyItem1 = new BuyItem();
-        buyItem1.setProductId(100);
-        buyItem1.setQuantity(1);
-        buyItemList.add(buyItem1);
-
-        createOrderRequest.setOrderList(buyItemList);
-        String json = objectMapper.writeValueAsString(createOrderRequest);
-        RequestBuilder requestBuilder = MockMvcRequestBuilders
-                .post("/users/{userId}/orders", 1)
-                .with(httpBasic("user2@gmail.com", "user2"))
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(json);
-
-        new CheckResult(mockMvc).check(requestBuilder, 403);
-    }
-
     // 查詢訂單列表
     @Test
     public void getOrders() throws Exception {
